[ROCm] Add async GMEM->LDS infra; switch KV tile load to inverse-swizzle addressing
Adds buffer_load_dword_lds primitives to memory_ops_hip.h: srsrc_t (V# buffer
descriptor), make_srsrc(), to_sgpr_u32(), and async_load_dword/tile64_to_lds().
The design follows the HipKittens global_to_shared pattern.  wait_group<N>() is
present but the s_waitcnt asm is commented out — async pipelining is not active
yet and will be a follow-up.
produce_kv and load_q_global_smem are rewritten to use inverse-swizzle
addressing: global memory is accessed via the swizzled index so LDS writes
use a simple linear offset.  Each call is split into a Phase A (all global
loads issued) and Phase B (all LDS stores), giving the hardware room to
pipeline the two without a double-buffer.  This restructuring is the
prerequisite for wiring in the async path.
permuted_smem.cuh extracts all swizzle arithmetic into a standalone
SwizzleLayout<mode> policy type.  smem_t is refactored to take it as a
template parameter (composition over inheritance), with the smem_t::Layout
alias serving as an explicit bijection handle between the global and LDS
index spaces.
Also fixes __umulhi to use v_mul_hi_u32 on gfx942 in fastdiv.cuh.

+++ include/gpu_iface/fastdiv.cuh
@@ -62,7 +62,7 @@ struct uint_fastdiv {
 #ifdef __CUDA_ARCH__
       q = __umulhi(m, n);
 #else
-      q = (((unsigned long long)((long long)m * (long long)n)) >> 32);
+      q = (uint32_t)(((uint64_t)m * (uint64_t)n) >> 32);
 #endif
       q += a * n;
       q >>= s;
@@ -80,7 +80,7 @@ __host__ __device__ __forceinline__ uint32_t operator/(const uint32_t n,
 #ifdef __CUDA_ARCH__
     q = __umulhi(divisor.m, n);
 #else
-    q = (((unsigned long long)((long long)divisor.m * (long long)n)) >> 32);
+    q = (uint32_t)(((uint64_t)divisor.m * (uint64_t)n) >> 32);
 #endif
     q += divisor.a * n;
     q >>= divisor.s;
